Remove unused code, refactor test code
- Remove unused sql jar and hbase access files
- Refactor hbase config into /config
- Change root log level to info for debugging
