+++ app/uk/gov/ons/sbr/models/units/EnterpriseUnit.scala
@@ -6,7 +6,6 @@ import io.swagger.annotations.ApiModelProperty
 import play.api.libs.json.{ JsValue, Json, OFormat }
 
 import uk.gov.ons.sbr.data.domain.Enterprise
-import uk.gov.ons.sbr.data.model.StatUnit
 import uk.gov.ons.sbr.models.DataUnit
 
 /**
@@ -30,12 +29,6 @@ object EnterpriseUnit {
     val childJson = o.getChildren.map {
       a => Json.parse(a.toUnitHierarchyAsJson)
     }.toList
-    // EnterpriseUnit(o.getKey.toLong, o.getReferencePeriod.toString, o.getVariables.toMap, o.getType.toString, childJson)
     EnterpriseUnit(o.getKey.toLong, o.getReferencePeriod.toString, o.getVariables.toMap, o.getType.toString, childJson)
   }
-
-  //  def apply(e: StatUnit): EnterpriseUnit = {
-  //    val childJson = e.children.map { x => Json.toJson(ChildUnit(x)) }.toList
-  //    EnterpriseUnit(e.key.toLong, e.refPeriod.toString, e.variables, e.unitType, childJson)
-  //  }
 }
