--- sbr-hbase-connector-1.0-SNAPSHOT-jar-with-dependencies-1.jar
@@ -1,1 +1,0 @@
-/tmp/hadoop-coolit/mapred/local/1517905063038/sbr-hbase-connector-1.0-SNAPSHOT-jar-with-dependencies-1.jar
